move some folders around
