+++ myft/ui/lib/myft-stuff-on-page-load.js
